[refactor] Moved vit encoder inside models/albef
