+++ mmf/models/__init__.py
@@ -16,6 +16,7 @@ from .fusions import FusionBase, ConcatBERT, ConcatBoW, LateFusion
 from .unimodal import UnimodalBase, UnimodalText, UnimodalModal
 from .visual_bert import VisualBERT
 from .vilbert import ViLBERT
+from .albef.vit import AlbefVitEncoder
 
 
 __all__ = [
@@ -41,4 +42,5 @@ __all__ = [
     "UnimodalBase",
     "UnimodalModal",
     "UnimodalText",
+    "AlbefVitEncoder"
 ]
