+++ mmf/models/albef/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (c) Facebook, Inc. and its affiliates.
+import mmf.models.albef.vit
